+++ addons/eu/en16931/en16931.go
@@ -128,5 +128,7 @@ func normalize(doc any) {
 		normalizeOrgIdentity(obj)
 	case *org.Inbox:
 		normalizeOrgInbox(obj)
+	case *org.Party:
+		normalizeOrgParty(obj)
 	}
 }